refactor: (22.02.24) Moved `SocketIoWorker` back to `workers` folder
